requirements.txt
cffi==1.14.0
click==7.1.1
cryptography==2.8
Flask==1.1.1
flask-lambda-support==0.1.5
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
pycparser==2.20
PyJWT==1.7.1
six==1.14.0
Werkzeug==1.0.0
